Typo in TableOfContents and other issues related to the Pixel Perfect doc update
The "Pixel Perfect extension" link in the TableOfContents.md file had a typo which linked to "CinemachineSPixelPerfect", this should have been "CinemachinePixelPerfect"

I left out the "CinemachineUsing.md" file in the last commit.

And... there were bad HREF links in the CinemachinePixelPerfect.md file. This included the wrong name for the 2D Pixel Perfect png file.